+++ src/main/java/com/flytrap/rssreader/api/account/domain/AccountId.java
@@ -1,8 +1,10 @@
 package com.flytrap.rssreader.api.account.domain;
 
+import java.io.Serializable;
+
 public record AccountId(
         long value
-) {
+) implements Serializable  {
     public AccountId {
         if (value < 0) {
             throw new IllegalArgumentException("Account id must be positive");
