--- src/main/java/com/flytrap/rssreader/api/member/domain/AccountId.java
@@ -1,13 +1,0 @@
-package com.flytrap.rssreader.api.member.domain;
-
-import com.flytrap.rssreader.global.model.DomainId;
-
-public record AccountId(
-    long value
-) implements DomainId {
-    public AccountId {
-        if (value < 0) {
-            throw new IllegalArgumentException("Account id must be positive");
-        }
-    }
-}
