+++ src/main/java/com/flytrap/rssreader/api/post/business/event/PostOpenEvent.java
@@ -1,6 +1,6 @@
 package com.flytrap.rssreader.api.post.business.event;
 
-import com.flytrap.rssreader.api.member.domain.AccountId;
+import com.flytrap.rssreader.api.account.domain.AccountId;
 import com.flytrap.rssreader.api.post.domain.PostAggregate;
 
 public record PostOpenEvent(
