+++ src/main/java/com/flytrap/rssreader/api/post/business/service/PostUpdateService.java
@@ -1,6 +1,6 @@
 package com.flytrap.rssreader.api.post.business.service;
 
-import com.flytrap.rssreader.api.member.domain.AccountId;
+import com.flytrap.rssreader.api.account.domain.AccountId;
 import com.flytrap.rssreader.api.post.domain.PostAggregate;
 import com.flytrap.rssreader.api.post.domain.PostId;
 import com.flytrap.rssreader.api.post.infrastructure.implementation.PostCommand;
