+++ src/main/java/com/flytrap/rssreader/api/post/infrastructure/implementation/PostQuery.java
@@ -1,7 +1,7 @@
 package com.flytrap.rssreader.api.post.infrastructure.implementation;
 
+import com.flytrap.rssreader.api.account.domain.AccountId;
 import com.flytrap.rssreader.api.folder.domain.FolderId;
-import com.flytrap.rssreader.api.member.domain.AccountId;
 import com.flytrap.rssreader.api.post.domain.Bookmark;
 import com.flytrap.rssreader.api.post.domain.Open;
 import com.flytrap.rssreader.api.post.domain.Post;
